+++ src/lib/queueContractInfoJobs.ts
@@ -1,11 +1,9 @@
 import SQS from 'aws-sdk/clients/sqs'
 import Bluebird from 'bluebird'
 import { chunk, orderBy, uniqBy } from 'lodash'
-import { Op } from 'sequelize'
 import { CONTRACT_INFO_QUEUE_URL } from './constants'
 import Logger from './logger'
 import stats from './stats'
-import { ContractInfo } from '@models/index'
 import { ProcessingGoal } from '@types'
 
 const logger = Logger(module)
@@ -29,43 +27,6 @@ export type TBackfill = {
   howMany?: number
   unproccessedOnly?: boolean
   minId?: number
-}
-
-export async function queueContractInfoBackfill({
-  howMany,
-  unproccessedOnly = true,
-  minId
-}: TBackfill) {
-  const whereClause =
-    minId || unproccessedOnly
-      ? {
-          ...(minId
-            ? {
-                id: {
-                  [Op.gte]: minId
-                }
-              }
-            : {}),
-          ...(unproccessedOnly
-            ? {
-                lastTransactionBlock: { [Op.is]: null }
-              }
-            : undefined)
-        }
-      : undefined
-
-  const contracts = await ContractInfo.findAll({
-    where: whereClause,
-    limit: howMany ? +howMany : undefined,
-    order: [['id', 'asc']]
-  })
-
-  const queueTokens: IContractInfoJobDetailsByTokenAddress[] = contracts.map(
-    contract => ({
-      tokenAddress: contract.address
-    })
-  )
-  await queueContractInfoByTokenAddressJobs(queueTokens)
 }
 
 /**
